requirements.txt
streamlit==1.38.0
opencv-python==4.10.0.84
easyocr==1.7.1
pandas==2.2.3
Pillow==10.4.0
numpy==1.26.4
tqdm==4.66.5
